+++ R/providers__oidc_discovery.R
@@ -438,6 +438,7 @@ oauth_provider_oidc_discover <- function(
     mtls_endpoint_aliases = mtls_endpoint_aliases,
     mtls_client_certificate_bound_access_tokens = tls_client_certificate_bound_access_tokens,
     issuer = iss,
+    oidc = TRUE,
     issuer_match = issuer_match,
     use_nonce = use_nonce,
     id_token_validation = id_token_validation,
@@ -454,6 +455,8 @@ oauth_provider_oidc_discover <- function(
   if (length(duplicate_dot_names) > 0) {
     provider_args[duplicate_dot_names] <- NULL
   }
+  provider_args[["oidc"]] <- TRUE
+  dots[["oidc"]] <- NULL
 
   do.call(oauth_provider, c(provider_args, dots))
 }